# backend/internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: AKK Portal Backend API
  version: "1.0.0"
  description: |
    API клиентского портала АО «Аграрная кредитная корпорация» (АКК).
    Единая точка для веб-портала (`web/`) и мобильного приложения.

    Контур: SMS-OTP авторизация, демо-вход eGov/Байтерек, подача кредитной
    заявки, движение по этапам (Смарт30-лестница), документы по этапам.

    **Авторизация.** Все эндпоинты `/api/v1/credit/*` и `/me` требуют
    `Authorization: Bearer <accessToken>`. Токен выдаётся при входе
    (`VerifySmsCode` с `purpose=login`), регистрации (`smsRegister`) или
    демо-SSO (`ssoDemoLogin`/`ssoEgovLogin`).

    **Demo-режим (`DEMO_MODE=true`, включён по умолчанию).** Реальная SMS не
    шлётся: код возвращается в поле `demoCode`. Демо-SSO eGov/Байтерек создаёт
    фиксированных клиентов и сразу выдаёт токены.
servers:
  - url: https://akk-backend-production.up.railway.app
    description: Production (Railway)
  - url: http://localhost:8080
    description: Local dev (PORT по умолчанию 8080)
...
paths:
...
  /api/v1/credit/applications:
    post:
      tags: [credit]
      summary: Создать кредитную заявку
      description: |
        Личность берётся из токена. Данные воронки клиента сохраняются в `onboarding`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requested_amount, loan_purpose]
              properties:
                requested_amount:
                  type: number
                  description: Запрошенная сумма, ₸
                  example: 5000000
                loan_purpose:
                  type: string
                  example: "Пополнение оборотных средств"
                program_id:
                  type: string
                  description: id выбранной программы (см. web data/programs)
                  example: "agrobusiness"
                onboarding:
                  type: object
                  description: Снимок воронки (ответы квиза, программа, параметры, контакт, стресс)
                  additionalProperties: true
      responses:
        "200":
          description: Заявка создана
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    get:
      tags: [credit]
      summary: Список заявок текущего клиента (новые сверху)
      responses:
        "200":
          description: Массив заявок
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Application" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/credit/applications/{uid}/status:
    get:
      tags: [credit]
      summary: Статус заявки + флаги для трекера и кнопки отмены
      parameters:
        - $ref: "#/components/parameters/AppUid"
      responses:
        "200":
          description: Статус заявки
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/credit/applications/{uid}/advance:
    post:
      tags: [credit]
      summary: Демо-движение заявки по этапам
      description: |
        Без тела — переход на следующий этап лестницы. Со `status` — установка
        конкретного (`new` — сброс, `rejected` — отказ). В прод движением управляет
        реальный процесс; здесь — ручное управление для демонстрации.
      parameters:
        - $ref: "#/components/parameters/AppUid"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: Целевой статус (этап лестницы / rejected / cancelled)
      responses:
        "200":
          description: Обновлённая заявка
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/credit/applications/{uid}/cancel:
    post:
      tags: [credit]
      summary: Самостоятельная отмена заявки заёмщиком
      description: Доступна только до решения кредитного комитета (new/scoring/expertise).
      parameters:
        - $ref: "#/components/parameters/AppUid"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, example: "Передумал" }
      responses:
        "200":
          description: Заявка отменена (status=cancelled)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: На текущем этапе отмена недоступна
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/credit/applications/{uid}/documents:
    get:
      tags: [credit]
      summary: Чек-лист документов заявки, сгруппированный по этапам
      parameters:
        - $ref: "#/components/parameters/AppUid"
      responses:
        "200":
          description: Документы по этапам
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentsDTO" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [credit]
      summary: Отметить требование как загруженное/подписанное (метаданные)
      description: Файл не передаётся — сохраняются только метаданные действия.
      parameters:
        - $ref: "#/components/parameters/AppUid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requirement_key]
              properties:
                requirement_key:
                  type: string
                  description: Ключ требования из каталога (см. /documents)
                  example: "business_plan"
                file_name: { type: string, example: "biznes-plan.pdf" }
      responses:
        "200":
          description: Документ отмечен
          content:
            application/json:
              schema:
                type: object
                properties:
                  requirement_key: { type: string }
                  status: { type: string, example: uploaded }
                  file_name: { type: string, nullable: true }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/credit/calculator:
    post:
      tags: [credit]
      summary: Калькулятор платежа (публичный pre-screen)
      description: |
        Расчёт графика 1-в-1 с веб-калькулятором (до тенге). Сумма клампится в
        [100000, max_amount]; срок для biannual_winter = effectiveMaxTerm (фикс),
        для annual — заданный term (или авто), не выше потолка цели.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [program_id, amount]
              properties:
                program_id: { type: string, example: agrobusiness }
                amount: { type: number, example: 10000000 }
                term:
                  type: integer
                  description: Срок, мес (для annual). Для biannual игнорируется.
                  example: 24
                purpose:
                  type: string
                  description: Цель — влияет на потолок срока (term_by_purpose)
                  example: investments
      responses:
        "200":
          description: Результат расчёта
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CalculatorResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: Программа не найдена
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/credit/notifications:
    get:
      tags: [credit]
      summary: Лента уведомлений личного кабинета
      description: |
        Строится из статусов заявок клиента (1-в-1 с вебом). `code` — для
        локализации на мобайле; `title`/`text` — готовые RU-строки.
      responses:
        "200":
          description: Уведомления + счётчик непрочитанных
          content:
            application/json:
              schema:
                type: object
                properties:
                  unread: { type: integer, description: Счётчик для бейджа }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Notification" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/credit/my-documents:
    get:
      tags: [credit]
      summary: Личное хранилище «Мои документы» (переиспользуемые + сроки действия)
      description: |
        Каталог переиспользуемых типов документов клиента со статусом по сроку:
        missing (нет) / valid / expiring (истекает ≤7 дн) / expired (просрочен).
        Валидные документы автоматически закрывают требования новых заявок —
        собирать одну и ту же стопку каждый раз не нужно.
      responses:
        "200":
          description: Массив типов с состоянием хранилища
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MyDocument" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [credit]
      summary: Сохранить/обновить документ в хранилище
      description: |
        `valid_until` вычисляется по сроку годности типа от `issued_at`
        (бессрочные типы → без срока). Метаданные (файл хранится отдельно).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [doc_type]
              properties:
                doc_type: { type: string, example: tax_clearance }
                file_name: { type: string, example: "spravka-kgd.pdf" }
                issued_at:
                  type: string
                  description: "Дата выдачи YYYY-MM-DD (пусто = сегодня)"
                  example: "2026-06-20"
      responses:
        "200":
          description: Сохранено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MyDocument" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AppUid:
      name: uid
      in: path
      required: true
      description: UUID заявки
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Требуется авторизация / недействительный токен
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
...
    Application:
      type: object
      properties:
        uid: { type: string, format: uuid }
        number: { type: string, example: "AKK-2026-000123" }
        program_id: { type: string }
        program: { type: string, description: Алиас program_id }
        purpose: { type: string }
        amount: { type: number, description: Сумма, ₸ }
        status:
          type: string
          description: Статус-этап лестницы (см. ApplicationStatus.workflow_status)
        onboarding:
          type: object
          description: Снимок воронки клиента
          additionalProperties: true
        created_at: { type: string, format: date-time }
    ApplicationStatus:
      type: object
      properties:
        uid: { type: string, format: uuid }
        workflow_status:
          type: string
          description: |
            Текущий этап. Лестница: new → scoring_in_progress → expertise →
            cc_approved → collateral_valuation → contract_signing → disbursed →
            monitoring → completed. Ветки: rejected, cancelled.
          example: scoring_in_progress
        is_terminal:
          type: boolean
          description: true → конечное состояние (completed/rejected/cancelled)
        can_cancel:
          type: boolean
          description: true → заёмщик может отменить (до решения КК)
        available_actions:
          type: array
          items: { type: string }
          description: Доступные действия (например ["cancel"])
    DocumentsDTO:
      type: object
      properties:
        application_uid: { type: string, format: uuid }
        current_status: { type: string }
        current_stage_index: { type: integer }
        stages:
          type: array
          items: { $ref: "#/components/schemas/DocStage" }
    DocStage:
      type: object
      properties:
        status_key: { type: string, example: expertise }
        stage_index: { type: integer }
        label: { type: string, example: "На рассмотрении" }
        documents:
          type: array
          items: { $ref: "#/components/schemas/AppDocument" }
    AppDocument:
      type: object
      properties:
        requirement_key: { type: string, example: business_plan }
        title: { type: string, example: "Бизнес-план / проектная смета" }
        source:
          type: string
          enum: [gov, upload, sign]
          description: gov — из госисточников (verified); upload — загрузка; sign — подпись ЭЦП
        status:
          type: string
          description: required | verified | uploaded
        file_name: { type: string, nullable: true }
        from_vault:
          type: boolean
          description: true → требование закрыто валидным документом из «Моих документов»
        needs_refresh:
          type: boolean
          description: true → документ в хранилище просрочен, нужно обновить
        valid_until:
          type: string
          description: "Срок действия из хранилища YYYY-MM-DD (если есть)"
    MyDocument:
      type: object
      properties:
        key: { type: string, example: tax_clearance }
        title: { type: string, example: "Справка об отсутствии налоговой задолженности (КГД)" }
        source:
          type: string
          enum: [gov, upload, sign]
        validity_days:
          type: integer
          description: "Срок годности, дней (0 = бессрочно)"
        reusable: { type: boolean }
        status:
          type: string
          enum: [missing, valid, expiring, expired]
        file_name: { type: string }
        issued_at: { type: string, description: "YYYY-MM-DD" }
        valid_until: { type: string, description: "YYYY-MM-DD (нет для бессрочных)" }
...
    CalculatorResult:
      type: object
      properties:
        program_id: { type: string }
        rate: { type: number }
        amount: { type: number, description: "Сумма после клампа [100000, max_amount]" }
        term: { type: integer, description: "Эффективный срок, мес" }
        schedule: { $ref: "#/components/schemas/Schedule" }
...
    Notification:
      type: object
      properties:
        kind:
          type: string
          enum: [ok, info, danger]
        code:
          type: string
          description: Машинный код для i18n
          enum: [application_accepted, application_stage, application_rejected, profile_ok]
        application_number: { type: string, example: "AKK-2026-000123" }
        stage_index: { type: integer, description: Индекс этапа (для application_stage) }
        title: { type: string, description: Готовая RU-строка }
        text: { type: string, description: Готовая RU-строка }
security:
  - bearerAuth: []